Password Criterion Amount Calculator
Criterion:
It is a six-digit number.
The value is within the range given in your puzzle input.
Going from left to right, the digits never decrease; they only ever increase or stay the same (like 111123 or 135679).
Two adjacent digits must be the same (more than two is counted as a fit if isAnyMultipleDouble is set to true)